Refactor: libstonithd: separate action-related code into own source file
Everything related to stonith_action_t has been moved from st_client.c to a new
st_actions.c, since st_client.c was ridiculously large, and the action stuff
isn't all client-related. No code was changed.

Before:
   2804 st_client.c

After:
   545 lib/fencing/st_actions.c
  2278 lib/fencing/st_client.c
