+++ lib/fencing/Makefile.am
@@ -22,7 +22,7 @@ libstonithd_la_LDFLAGS	+= $(LDFLAGS_HARDENED_LIB)
 libstonithd_la_LIBADD	= $(top_builddir)/lib/common/libcrmcommon.la
 libstonithd_la_LIBADD   += $(top_builddir)/lib/services/libcrmservice.la
 
-libstonithd_la_SOURCES	= st_client.c st_output.c st_rhcs.c
+libstonithd_la_SOURCES	= st_actions.c st_client.c st_output.c st_rhcs.c
 if BUILD_LHA_SUPPORT
 libstonithd_la_SOURCES	+= st_lha.c
 endif
